+++ react/components/NestedExtensionPoints.tsx
@@ -1,7 +1,7 @@
 import {canUseDOM} from 'exenv'
 import PropTypes from 'prop-types'
 import React, {PureComponent} from 'react'
-import MaybeAuth from './MaybeAuth'
+import Loading from './Loading'
 
 import ExtensionPoint from '../ExtensionPoint'
 import {getPagePath, getParams} from '../utils/pages'
@@ -11,7 +11,8 @@ const EMPTY_OBJECT = {}
 
 interface Props {
   page: string,
-  query?: Record<string, string>
+  query?: Record<string, string>,
+  breakPoint?: Record<string, any>
 }
 
 export default class NestedExtensionPoints extends PureComponent<Props> {
@@ -35,16 +36,30 @@ export default class NestedExtensionPoints extends PureComponent<Props> {
     // a/b/c should render three extension points
     // <a><b><c></c></b></a>
     const getNestedExtensionPoints = (runtime: RenderContext) => {
-      return reverse.reduce((acc: JSX.Element | null, value: string, index: number) => (
-        <MaybeAuth pages={runtime.pages} page={page} navigate={runtime.navigate} segment={value}>
+      return reverse.reduce((acc: JSX.Element | null, value: string, index: number) => {
+        const { breakPoint } = this.props
+        if (breakPoint && breakPoint.point && breakPoint.point === value) {
+          if (breakPoint.loading) {
+            return <div className="flex justify-center ma4"><Loading /></div>
+          } else if (breakPoint.logged) {
+            return <ExtensionPoint
+              id={value}
+              query={query}
+              params={this.getPageParams(runtime, segments.slice(0, segments.length - index).join('/'))}>
+              {acc}
+            </ExtensionPoint>
+          }
+          return null
+        }
+        return (
           <ExtensionPoint
             id={value}
             query={query}
             params={this.getPageParams(runtime, segments.slice(0, segments.length - index).join('/'))}>
             {acc}
           </ExtensionPoint>
-        </MaybeAuth>
-      ), null as JSX.Element | null)
+        )
+      }, null as JSX.Element | null)
     }
 
     return <RenderContext.Consumer>{getNestedExtensionPoints}</RenderContext.Consumer>